agents_sdk/requirements.txt
openai>=1.0.0,<2.0.0
python-dotenv>=0.19.0
openai-agents>=0.1.0
pathlib>=1.0.1 